+++ src/protocol/index.ts
@@ -3,6 +3,7 @@ export * from "./cc";
 export * from "./delay-clock-rate";
 export * from "./enums";
 export * from "./lfo-clock-rate";
+export * from "./preset";
 export * from "./transpose";
 export { Tune } from "./tune";
 export { Voices } from "./voices";